# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Algo Trader API
  description: Algorithmic trading platform API
  version: 3.1.9
  license: { name: Proprietary }
servers:
  - url: http://localhost:8787
    description: Local dev
  - url: https://algo-trader.workers.dev
    description: Production
security: [{ BearerAuth: [] }]
...
paths:
...
  /api/v1/risk/var:
    post:
      tags: [Risk]
      summary: VaR/CVaR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [positions]
              properties: { positions: { type: array } }
      responses:
        '200':
          description: VaR
          content: { application/json: { schema: { $ref: '#/components/schemas/VaRResult' } } }
...
  /api/v1/risk/atr/stop:
    post:
      tags: [Risk]
      summary: ATR stop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol, exchange, entryPrice]
              properties:
                symbol: { type: string }
                exchange: { type: string }
                entryPrice: { type: number }
      responses:
        '200':
          description: Stop
          content:
            application/json:
              schema: { type: object, properties: { stopPrice: { type: number } } }
  /api/v1/risk/kelly/size:
    post:
      tags: [Risk]
      summary: Kelly sizing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [winRate, avgWin, avgLoss]
              properties:
                winRate: { type: number }
                avgWin: { type: number }
                avgLoss: { type: number }
      responses:
        '200':
          description: Fraction
          content:
            application/json:
              schema: { type: object, properties: { kellyFraction: { type: number } } }
...
  /api/v1/risk/correlation:
    post:
      tags: [Risk]
      summary: Correlation matrix
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbols]
              properties:
                symbols: { type: array, items: { type: string } }
                period: { type: integer, default: 30 }
      responses:
        '200':
          description: Correlation matrix
        '500':
          description: Internal server error
  /api/v1/risk/drawdown/alert:
    post:
      tags: [Risk]
      summary: Trigger drawdown alert check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                threshold: { type: number }
      responses:
        '200':
          description: Drawdown alert result
        '500':
          description: Internal server error
  /api/v1/risk/atr/stop/{symbol}:
    get:
      tags: [Risk]
      summary: Get stored ATR state for symbol
      parameters:
        - { name: symbol, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: ATR state
        '404':
          description: No ATR state found for symbol
        '500':
          description: Internal server error
    delete:
      tags: [Risk]
      summary: Clear ATR state (position close)
      parameters:
        - { name: symbol, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: ATR state cleared
        '500':
          description: Internal server error
  /api/v1/risk/kelly/from-history:
    post:
      tags: [Risk]
      summary: Kelly sizing from trade history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trades]
              properties:
                trades: { type: array }
                bankroll: { type: number }
      responses:
        '200':
          description: Kelly result
        '500':
          description: Internal server error
  /api/v1/risk/cache:
    delete:
      tags: [Risk]
      summary: Invalidate user risk cache
      responses:
        '200':
          description: Cache invalidated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
        '500':
          description: Internal server error

  # ── Referral (additional sub-routes) ───────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
  schemas:
...
    VaRResult:
      type: object
      properties:
        var: { type: number }
        cvar: { type: number }
...